added -Wall -Werror
From: Alexander Petry <alexander.petry@itwm.fraunhofer.de>

git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@398 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/fhglog/trunk/fhglog/LogLevel.cpp
@@ -34,6 +34,6 @@ const std::string &LogLevel::str() const
     , "FATAL"
     , "UNSET"
   };
-  assert(lvl_ >= 0 && lvl_ < sizeof(LevelToStringMap_));
+  assert(lvl_ >= 0 && lvl_ < static_cast<int>(sizeof(LevelToStringMap_)));
   return LevelToStringMap_[lvl_];
 }
